feat: refactor NewsItem parsing into root g2 package
- Moved `NewsItem` model to `g2/news.go`.
- Moved parsing logic and `StripEmail` to `g2/news_parser.go`.
- Added serialization methods `ToHTMLTemplate()`, `ToHTML()`, and `ToText()` to `NewsItem` for easier reconstruction.
- Cleaned up `cmd/g2/site.go` to use the new root-level model and removed local implementations.
- Removed `plan.md`.
