+++ cmd/g2/sitegen_templates/news_article.html
@@ -16,5 +16,5 @@
         {{end}}
     </div>
 
-    <div class="news-body" style="margin-top: 1em;">{{.NewsItem.BodyHTML}}</div>
+    <div class="news-body" style="margin-top: 1em;">{{.NewsItem.ToHTMLTemplate}}</div>
 </article>
